refactor: 💡 Restructure code a bit
